Fix backwards compatibility mechanism
